Ldaptive v2 grouper v5

+++ grouper/src/grouper/edu/internet2/middleware/grouper/app/ldapProvisioning/ldapSyncDao/LdapSyncDao.java
@@ -52,7 +52,7 @@ public abstract class LdapSyncDao {
    * @param sizeLimit
    * @return the data
    */
-  public abstract List<LdapEntry> search(String ldapPoolName, String baseDn, String filter, LdapSearchScope ldapSearchScope, List<String> attributeNames, Long sizeLimit);
+  public abstract List<LdapEntry> search(String ldapPoolName, String baseDn, String filter, LdapSearchScope ldapSearchScope, List<String> attributeNames, Integer sizeLimit);
   
   
   /**
